# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi System Laragpt",
        "description": "Descripción de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/challenges": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Show All Challenges",
                "operationId": "df61bd7503e62a2251f358931e0f3a5f",
                "responses": {
                    "200": {
                        "description": "List of Challenges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "My title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "My Description"
                                                    },
                                                    "difficulty": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/challenges?page=1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/challenges?page=3"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/challenges?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "24"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Almacena un nuevo challenge en el sistema.",
                "operationId": "63ffd18fb5f5371ec322250145cbb046",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "difficulty",
                                    "user_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "My title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "My Description"
                                    },
                                    "difficulty": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Entered Successfully"
                                        },
                                        "last_insert_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/challenges/{id}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Show Specific Challenge",
                "operationId": "ed8928d8cd2720e0dcce6cd0e29767f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "My title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "My Description"
                                        },
                                        "difficulty": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No exists challenge with id : #"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/challenges/{challenge}": {
            "put": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Update Exist Challenge in Database",
                "operationId": "a7c91e857e50cd280fbafc18ad9f6df5",
                "parameters": [
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "ID del challenge a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "difficulty",
                                    "user_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "My title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "My Description"
                                    },
                                    "difficulty": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Update Successfully"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge no encontrado"
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Remove Challenge From System",
                "operationId": "ee00796c790a08f161efe1bb28f0773b",
                "parameters": [
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "ID del challenge a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Delete Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge no encontrado"
                    }
                }
            }
        },
...
    }
}